Support orientation handling
Original support code always put the plates across the shorter dimension. This update modifies the default algorithm to go lengthwise if the support area is too narrow. It also adds a new "orient" parameter that can be used to force plates to go lengthwise. The expanded call is backwardly compatible and the orient defaults to the short dimension.